Version: .03
Enemy Movement: The Enemy will move on the Navmesh however adding to many Navmesh surfaces will cause the enemy to jitter around. My guess is the enemy can not figure out what path to take. The enemy has a set target range so if the player is in this range the enemy will start moving to the player.

Enemy Spawn: We have the ability to spawn multiple random enemies in the game. The enemy will spawn based around the players position which can be changed in the inspector.