Player Input
When you play the game from the main menu the Game Master will add the Player Input component and it's settings. However if you start from the Game scene Player Input doesn't exist. So in the game scene I added the Player Input componet to the Game Master object and somehow that fixed it.