Game Scene and collsion updates
Ok so this is what I got. The navmesh has all the jump points it needs to function for the enemy to jump. I got it working.
Collsion with Enemy to player works to will players health if the enemy collides with the player.
Sowrd should be able to deal damage to the enemy. The only problem is besides moving none of the key binds seem to work correctly.
Aim Camera and bow. The aim camera will follower the players rotation. However for the bow object I was unable to apply rotation to the bow. This is because if the Bow Object is a parent of any Players game object (For example Player, PlayerObj, mixmo stuff) the bows y rotation is automatically set at -90 and changing it to zero would cause the bow to point down. This is not the only problem. On a unparented bow object pitch (up and down) is the x Axis and yaw (right and left) is the y axis. But once you parent the bow to PlayerObj pitch becomes the z axis but yaw stays the same. These many factors make it very confusing to mange and track the bows rotation.
and by difficult I mean almost impossible.